refactor(metabox-translations): Extract common code to both editors
